--- package.json
@@ -1,5 +1,0 @@
-{
-  "dependencies": {
-    "ws": "^8.17.1"
-  }
-}
