--- zooming/index.html
@@ -1,12 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-    <title>Catan</title>
-    <link rel="stylesheet" type="text/css" href="style.css">
-</head>
-<body>
-    <svg id="test" width="100" height="100" style="left: 0px; top: 0px">
-    </svg>
-    <script src="script.js"></script>
-</body>
-</html>
