--- zooming/style.css
@@ -1,12 +1,0 @@
-html, body {
-    margin: 0;
-    padding: 0;
-    overflow: hidden;
-    background-color: #004080;
-    width: 100vw;
-    height: 100vh;
-}
-svg {
-    position: absolute;
-    border: 3px solid black;
-}
